Don't set a shell/home dir for swarm client user
Since we are now using the same user ID for deployment and running of
the Jenkins Swarm client service, we need to make sure that it's a
valid user that can log in to the machine.

Regarding the home directory, it seems that this value cannot be set
by Ansible if the same user is logged into the machine via ssh at the
time.