Add restart task in jenkins-swarm-client role
Because the shutdown and startup tasks rely on a number of variables
defined by this role, it is not feasible to include_tasks for
shutdown/startup directly. This task allows us to restart the service
via an include_role task with this file given as in the tasks_from
argument.