Run Jenkins Swarm Client as a LaunchAgent
According to Apple's documentation on Daemons and Agents:

The difference between an agent and a daemon is that an agent can
display GUI if it wants to, while a daemon can't. The difference
between an agent and a regular application is that an agent typically
displays no GUI (or a very limited GUI).

This change, along with modifying the base template image to
automatically log in the Jenkins user, will allow the swarm client to
run processes that want to attach to the WindowServer.