Restart swarm client service on-failure
The on-abort argument to Restart= only will restart the service if it
exits with an unclean signal. If the service crashes or dies for other
reasons, it will not be restarted. This was recently observed in
action when something caused the JVM to crash, which exits with a
"clean" signal but a failure code.

This change uses the on-failure value instead, which will restart the
service not only on unclean signal codes, but also unclean exit codes,
watchdog errors, and timeouts. For more information, see:

https://www.freedesktop.org/software/systemd/man/systemd.service.html#Restart=