Use file leak detector on Windows nodes (for now)
This is an attempt to understand why sometimes Windows jobs fail to
wipe the workspace when cloning. This phenomenon is observed on crate
jobs, though given that they are currently doing the most Windows
builds, I believe that this is not exclusive to their project but
perhaps part of a wider but unexposed phenomenon.

At any rate, the leak detector will start an HTTP service on
http://localhost:1999, which shows a listing of all open filehandles
by the given Java process and all subprocesses. It is our hope that
the next time we see failed git checkouts on Windows, we can use this
resource to see which process is preventing the workspace from being
deleted.

For more information, see https://file-leak-detector.kohsuke.org/