Remove leak detector JAR for Windows
We haven't seen build cleanup failures on Windows in ages. Even when
we did, this JAR was not very useful in diagnosing the problem, since
it never found a "smoking gun". Furthermore, something changed on
Maven's central repo in terms of how they deliver JAR files from
search.maven.org, and this now causes a failure in Ansible.

For all of these reasons, it's better to just remove the leak detector
JAR for this platform.