Load of launchd-based services more safely
This change fixes an issue where an invalid plist file for a launchd
service would result these tasks reporting success, but the system
would have never started the service.

Part of the problem here is that launchctl does not have a great
interface as far as CLI tools go, and while the documentation warns
against parsing the text output of the tool, we don't have much choice
here. The launchctl tool also does not return non-zero integer codes
when something fails, which requires us to grep the text output.

This change always loads the launchd service, but now will test
afterwards to see if the service is running. This test will be skipped
if the service is already loaded.