Assemble a list of arguments for the swarm client
This commit fixes an issue introduced in 37e6642, which could cause
the swarm client argument list to contain empty (whitespace-only)
elements. This, in turn, caused problems for launchctl, since it does
not like empty element pairs.

This change also removes the java_exe variable from the list of
command arguments, which always felt a bit weird.