Set ignore_errors on task, not failed_when
Because the "which python3" task had the "failed_when: false"
attribute, it would always return a status of "ok", and hence, we
couldn't use "is success" in the follow-up task's `when` condition.
Using "ignore_errors: true" is a better way to check the output of the
command. Should have been done in e5c77c78.