Wait 60 seconds before restaring Swarm Client
We are seeing occasional crashes from Swarm Client, which may be
related to network flakiness, though this is not confirmed. However,
we do know that the service is being restarted too quickly, since
looking at the journal for this service shows the log filled with
"Start request repeated too quickly" messages before systemd gives up.

The default timeout is 100ms, which is far too fast in case of network
flakiness. Also see:

https://www.freedesktop.org/software/systemd/man/systemd.service.html#RestartSec=