Set ProcessType to Interactive for Mac service
According to Apple's documentation:

"Interactive jobs run with the same resource limitations as apps, that
is to say, none. Interactive jobs are critical to maintaining a
responsive user experience, and this key should only be used if an
app's ability to be responsive depends on it, and cannot be made
Adaptive."

In this case, we want the service to have the same priority as a GUI
application, and this is also relevant in case GUI applications are
being run as a part of builds.