Add --build-from-source to <10.14 brew options
Recently, Apple and Homebrew cut off support for pre-Mojave versions
of macOS, which means that there will no longer be bottles (pre-built
packages), and everything must be built by hand. Some formulae may not
build, but we'll need to cross that bridge when we get there.

Regarding the duplication of the homebrew_install_options tasks, this
is necessary because at some point in the future, we will split these
roles off to separate repositories.