Load Swarm Client on Mac in Aqua login session
This change loads the Swarm Client service on macOS into the logged-in
user's session, which allows it to access its keychain and other
session-related data. As we are now running the service for the user,
the use of sudo/become is no longer necessary.

In the future, we may want to consider making this a configurable
option, as it may not be appropriate for all use-cases.