Support running Swarm Client as an autologin task
This is necessary because it seems to be the only way to start a
service which can interact with the Windows desktop. Sadly, the
startup task is a somewhat large hammer for this particular nail, but
there isn't really a more elegant way to do this, as the comment
explains.