Ignore errors when logon task service shuts down
If the PID file is stale (due to the service crashing etc), then this
task will fail and halt a deployment. This isn't a problem with NSSM,
since NSSM will keep the service alive in such cases, but we don't
have the same guarantee when starting the service with a logon task.