Remove quotes in swarm client's logging argument
On Mac/Windows, this argument doesn't get correctly parsed by Java,
which leads to logging not being enabled. On Linux it seems to work
fine (and I do recall that it also worked on Mac/Windows long ago). At
any rate, removing the quotes seems to work. I also tried moving the
quotes around the filename itself and not the whole argument, but that
didn't work either.

This change of course introduces the risk that something will break if
the Jenkins Swarm home has spaces in it (which it does not in our
case). This has also been documented to prevent potential problems.