Don't wait for network when starting service
I suspect that this is the reason why the service sometimes fails to
launch on Windows -- the scheduled task is set to fire once when the
user logs in, but the network isn't ready yet and the trigger does not
retry once the network is up.

If the swarm client is running and the network is not online yet, then
it will fail to connect to the host and enter a retry loop. Therefore it
should be safe to start the service without the network.