Replace most include_* tasks with import_*
The difference between include_* and import_* is that the former is
dynamic and the latter is static, meaning that include_* is processed
at runtime and import_* is processed during parsing of the playbook.
As such, import_* tasks cannot be used with filenames that include
interpolated variables which must to be processed at runtime.

Using import_* tasks is generally preferred, since it increases the
effectiveness of linters and similar tools. However, the catch is that
mixing import_* and include_* tasks does not always work.
Specifically, calling an import_* task from within an include_*
context will fail.

This commit converts *most* of our include_* tasks to import_* ones,
except for those which use variable filenames or are included within a
dynamic context.