Use sysvinit module when SysV is required
This helps to avoid problems with the `system` module misdetects the
service manager, which can happen when using OpenSSH inside of a
Docker container. This use-case might be a problem for us in the near
future if we revive the Docker environment.

Even if we don't revive the Docker environment, this is a sound fix
because we know that we'll be using SysVinit if we are in this file.