Use win_service module instead of win_shell
Sadly, we still need a separate task for this, as the win_nssm task
doesn't support setting the username. According to the documentation
of the `user` attribute:

DEPRECATED, will be removed in a major release after 2021-07-01,
please use the ansible.windows.win_service module instead.

As such, this commit does just that.