Support reading swarm version from plugins.yaml
A longstanding "gotcha" for this role (at least in our infrastructure)
is the fact that one needs to take care to ensure that the Swarm Client
controller/node version match. Usually this means that whenever we
update our plugins.yaml file, a corresponding change must also be made
to the vars file that contains the jenkins_swarm_client_version fact.
This is cumbersome and has resulted in accidental updates where the
server/client version were not aligned.

With this commit, this role now supports passing the plugins.yaml
version file to the role, and the Swarm Client version will be extracted
from that. This means a bit less data duplication and safer deployments.

It is still possible to specify the jenkins_swarm_client_version, which
is respected just as before. However, this fact cannot be used in
combination with jenkins_plugins_file, so as a result, the default value
(which is a very old and obsolete version of Swarm Client) has been
removed and set to the empty string. This is a breaking API change and
as such, the minor version will be bumped in the next commit.