Store the Swarm Client PID file in /tmp on Unix
In https://github.com/jenkinsci/swarm-plugin/pull/114, an unfortunate
feature was added to Swarm Client which makes it exit on startup if the
PID file already exists (presumably to prevent multiple instances from
being started). Unfortunate, because while this feature might be useful
to users who manually manage the service or use a primitive service
manager, that is not the case for this role.

This role uses service managers on Unix that already prevent multiple
instances of a process from being started. But, thanks to this feature,
if the system suddenly reboots or Swarm Client fails to exit cleanly,
then the PID file won't be removed and the service will refuse to start.

Ultimately, this behavior should be fixed in the Swarm Client itself,
but until then we should at least place the PID file in /tmp on Unix
hosts, since this directory is cleaned on boot. While Windows has both a
system and user temporary directory, they are not cleaned on boot and
therefore this change does not change the PID file location on that
platform.