Don't template jenkins_swarm_config_args if empty
The default value of `jenkins_swarm_config_args` is an empty dict
(`{}`), which `to_nice_yaml` doesn't mind and inserts literally into the
config file. Unfortunately, that's not valid YAML and the swarm client
chokes on it when trying to load such a value. This change only adds the
templated value if it is true-ish, which will not be the case for an
empty dict.