Decouple enabling/disabling and starting/stopping
The previous state was a bit of a mess, especially on Mac where we
called `enable-service.yml` from the `startup.yml` file, and in turn,
`enable-service.yml` called `startup.yml`. A condition luckily prevented
the recursive include from going off the rails.

This change decouples the starting/stopping and enabling/disabling tasks
so that each only does exactly that. As such, it's a breaking change and
we'll bump the minor version of this role in the next commit.